Random munging with the buffer status stuff

+++ vorbis-tools/ogg123/ogg123.h
@@ -50,7 +50,7 @@ void play_file(ogg123_options_t opt);
 int get_tcp_socket(void); /* Will be going soon. */
 FILE *http_open(char *server, int port, char *path); /* ditto */
 int open_audio_devices(ogg123_options_t *opt, int rate, int channels, buf_t ** buffer);
-double buffer_full (buf_t* buffer);
+long buffer_full (buf_t* buffer);
 void buffer_cleanup (void);
 void signal_quit (int ignored);
